Added mutations for form submission and user input to query heroes based on severity selected
