--- client/src/pages/stripe.js
@@ -1,19 +1,0 @@
-import { Elements } from '@stripe/react-stripe-js';
-import { loadStripe } from '@stripe/stripe-js';
-
-// Make sure to call `loadStripe` outside of a component’s render to avoid
-// recreating the `Stripe` object on every render.
-const stripePromise = loadStripe('pk_test_TYooMQauvdEDq54NiTphI7jx');
-
-function App() {
-    const options = {
-        // passing the client secret obtained from the server
-        clientSecret: '{{CLIENT_SECRET}}',
-    };
-
-    return (
-        <Elements stripe={stripePromise} options={options}>
-            <CheckoutForm />
-        </Elements>
-    );
-};
