+++ client/src/utils/helpers.js
@@ -1,5 +1,54 @@
 export function validateEmail(email) {
-    const re = /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
-    return re.test(String(email).toLowerCase());
-  }
-  
+  const re = /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
+  return re.test(String(email).toLowerCase());
+}
+
+
+export function idbPromise(donation, method, object) {
+  return new Promise((resolve, reject) => {
+    const request = window.indexedDB.open('donation-donation', 1);
+    let db, tx, store;
+    request.onupgradeneeded = function (e) {
+      const db = request.result;
+      db.createObjectStore('donation', { keyPath: '_id' });
+
+    };
+
+    request.onerror = function (e) {
+      console.log('There was an error');
+    };
+
+    request.onsuccess = function (e) {
+      db = request.result;
+      tx = db.transaction(donation, 'readwrite');
+      store = tx.objectStore(donation);
+
+      db.onerror = function (e) {
+        console.log('error', e);
+      };
+
+      switch (method) {
+        case 'put':
+          donation.put(object);
+          resolve(object);
+          break;
+        case 'get':
+          const all = store.getAll();
+          all.onsuccess = function () {
+            resolve(all.result);
+          };
+          break;
+        case 'delete':
+          store.delete(object._id);
+          break;
+        default:
+          console.log('No valid method');
+          break;
+      }
+
+      tx.oncomplete = function () {
+        db.close();
+      };
+    };
+  });
+}
