+++ package.json
@@ -31,6 +31,7 @@
     "jsonwebtoken": "^8.5.1",
     "jwt-decode": "^3.1.2",
     "react-bootstrap": "^2.1.1",
-    "react-icons": "^4.3.1"
+    "react-icons": "^4.3.1",
+    "stripe": "^8.200.0"
   }
 }
